% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCPexperiments.R
\name{LCPcompare0}
\alias{LCPcompare0}
\title{LCPcompare0 - A wrapper for comparing CP/LCP using regression
imports:}
\usage{
LCPcompare0(
  xtrain,
  ytrain,
  xcalibration,
  ycalibration,
  xtest,
  ytest,
  alpha = 0.05,
  quantiles = c(0.025, 0.05, 0.1, 0.9, 0.95, 0.975, ),
  nfolds = 3,
  random_state = 1,
  save_path = NULL,
  print_out = 10,
  epochs = 50
)
}
\description{
LCPcompare0 - A wrapper for comparing CP/LCP using regression
imports:
}
